fix some warnings
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@22266 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/prntbase.cpp
@@ -737,8 +737,8 @@ void wxPrintPreviewBase::AdjustScrollbars(wxPreviewCanvas *canvas)
     double actualHeight = (zoomScale*m_pageHeight*m_previewScale);
 
     // Set the scrollbars appropriately
-    int totalWidth = actualWidth + 2*m_leftMargin;
-    int totalHeight = actualHeight + 2*m_topMargin;
+    int totalWidth = (int)(actualWidth + 2*m_leftMargin);
+    int totalHeight = (int)(actualHeight + 2*m_topMargin);
     int scrollUnitsX = totalWidth/10;
     int scrollUnitsY = totalHeight/10;
     wxSize virtualSize = canvas->GetVirtualSize();
